0x0C-python-almost_a_circle/models/rectangle.py
#!/usr/bin/python3
"""
A module for a rectangle
"""
from models.base import Base


class Rectangle(Base):
    """
    A class that defines a rectangle
    """

    def __init__(self, width, height, x=0, y=0, id=None):
        """
        Initializes an instance of Rectangle
        """
        super().__init__(id)
        self.__width = width
        self.__height = height
        self.__x = x
        self.__y = y

    @property
    def width(self):
        """
        returns the width
        """
        return self.__width

    @width.setter
    def width(self, value):
        """
        sets the width
        """
        self.__width = value

    @property
    def height(self):
        """
        returns the height
        """
        return self.__height

    @height.setter
    def height(self, value):
        """
        sets the value of height
        """
        self.__height = value

    @property
    def x(self):
        """
        returns the value of x
        """
        return self.__x

    @x.setter
    def x(self, value):
        """
        sets the value of x
        """
        self.__x = value

    @property
    def y(self):
        """
        returns the value of y
        """
        return self.__y

    @y.setter
    def y(self, value):
        """
        sets the value of y
        """
        self.__y = value
